Fixed websocket code for 32 bit OS

+++ data/flowlang/mwtg/hp18/2164/edb0/mwtghp182164edb0bk19.rs
@@ -1,6 +1,6 @@
 let mut msg = msg.as_bytes();
 
-let n = msg.len();
+let n = msg.len() as i64;
 let mut reply: Vec<u8> = Vec::new();
 
 reply.push(129); // Text = 129 / Binary = 130;
